Improved full and partial skeleton creation testing

+++ Straight_skeleton_2/include/CGAL/constructions/Straight_skeleton_cons_ftC2.h
@@ -49,7 +49,7 @@ inline CORE::BigFloat MP_Float_to_BigFloat( MP_Float const& b )
     d += d_exp * CORE::BigFloat(b.of_exp(i));
   }
 
-  return d * CORE::BigFloat::exp2(exp * log_limb);
+  return d * CORE::BigFloat::exp2( static_cast<int>(exp * log_limb) );
 }
 #endif
